put myProfile class in myPage
